pyproject.toml
[project]
name = "prism"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "uvicorn>=0.40.0",
]

[tool.uv.workspace]
members = [
    "trial",
]
